# api/swagger/swagger.yaml
definitions:
  github_com_deva-labs_univia_internal_api_modules_role_models.Role:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_deva-labs_univia_internal_api_modules_user_models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      googleID:
        type: string
      id:
        type: string
      password:
        type: string
      phoneNumber:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/github_com_deva-labs_univia_internal_api_modules_role_models.Role'
        description: Foreign key to Role
      roleID:
        type: string
      status:
        type: boolean
      twitterID:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.ChangePasswordRequest:
    properties:
      new_password:
        example: newPassword123
        minLength: 8
        type: string
      old_password:
        example: oldPassword123
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - new_password
    - old_password
    - user_id
    type: object
...
  github_com_deva-labs_univia_pkg_types.ForgotPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
  github_com_deva-labs_univia_pkg_types.GetUserAvatarRequest:
    properties:
      user_id:
        example: 1
        type: integer
    required:
    - user_id
    type: object
  github_com_deva-labs_univia_pkg_types.GoogleLoginRequest:
    properties:
      token:
        example: google-oauth-token
        type: string
    required:
    - token
    type: object
...
  github_com_deva-labs_univia_pkg_types.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
      phone_number:
        example: "+84901234567"
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    type: object
  github_com_deva-labs_univia_pkg_types.RenewPasswordRequest:
    properties:
      new_password:
        example: newPassword123
        minLength: 8
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - new_password
    - user_id
    type: object
...
  github_com_deva-labs_univia_pkg_types.StatusBadRequest:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad Request
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.StatusForbidden:
    properties:
      code:
        example: 403
        type: integer
      message:
        example: Forbidden
        type: string
    type: object
  github_com_deva-labs_univia_pkg_types.StatusInternalError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal Server Error
        type: string
    type: object
  github_com_deva-labs_univia_pkg_types.StatusOK:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Success
        type: string
    type: object
  github_com_deva-labs_univia_pkg_types.StatusResponseVerification:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Verification code sent to your email. Please verify to process.
        type: string
    type: object
  github_com_deva-labs_univia_pkg_types.StatusUnauthorized:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.SuccessLoginEmailResponse:
    properties:
      data:
        example: example@hotmail.com
        type: string
      status:
        $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusResponseVerification'
    type: object
  github_com_deva-labs_univia_pkg_types.SuccessResponse:
    properties:
      data: {}
      status:
        $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusOK'
    type: object
...
  github_com_deva-labs_univia_pkg_types.TwitterLoginRequest:
    properties:
      background_color:
        example: '#1DA1F2'
        type: string
      background_image:
        example: https://example.com/bg.jpg
        type: string
      email:
        example: user@example.com
        type: string
      image:
        example: https://example.com/avatar.jpg
        type: string
      twitter_id:
        example: "123456789"
        type: string
      username:
        example: johndoe
        type: string
    required:
    - email
    - twitter_id
    - username
    type: object
...
info:
  contact: {}
paths:
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change password by providing old and new password
      parameters:
      - description: Change password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      summary: Change Password
      tags:
      - Authentication
...
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a verification email for password reset
      parameters:
      - description: Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification email sent
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      summary: Forgot Password
      tags:
      - Authentication
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login using email, username or phone + password
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login success with verification code sent (data is email)
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessLoginEmailResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusUnauthorized'
      summary: Login with email/username/phone
      tags:
      - Authentication
  /api/v1/auth/login/google:
    post:
      consumes:
      - application/json
      description: Login using Google OAuth token
      parameters:
      - description: Google login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login success with tokens
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
      summary: Login with Google
      tags:
      - Authentication
  /api/v1/auth/login/twitter:
    post:
      consumes:
      - application/json
      description: Login using Twitter account info
      parameters:
      - description: Twitter login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.TwitterLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login success with tokens
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
      summary: Login with Twitter
      tags:
      - Authentication
  /api/v1/auth/refresh-access-token:
    post:
      consumes:
      - application/json
      description: Refresh JWT token with refresh token + client id
      parameters:
      - description: Refresh Token (Bearer token format)
        in: header
        name: x-rtoken-id
        required: true
        type: string
      - description: Client ID
        in: header
        name: x-client-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: New tokens
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      summary: Refresh Access Token
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with email/username/password
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_internal_api_modules_user_models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created user
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deva-labs_univia_internal_api_modules_user_models.User'
                status:
                  $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusOK'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      summary: Register new user
      tags:
      - Authentication
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password without old password (via verification flow)
      parameters:
      - description: Renew password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.RenewPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      summary: Renew Password
      tags:
      - Authentication
  /api/v1/auth/user-info:
    get:
      consumes:
      - application/json
      description: Retrieve the user information by given token
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User info
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deva-labs_univia_internal_api_modules_user_models.User'
                status:
                  $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusOK'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusUnauthorized'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusForbidden'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      security:
      - BearerAuth: []
      summary: Get User Info
      tags:
      - User
...
  /api/v1/users/avatar:
    post:
      consumes:
      - application/json
      description: Get avatar of a user by user_id
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.GetUserAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Avatar URL
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      security:
      - BearerAuth: []
      summary: Get User Avatar
      tags:
      - User
...
swagger: "2.0"