# internal/api/modules/role/models/role.model.go
package roles

import (
	"time"

	"github.com/google/uuid"
)

// Role represents the role model.
type Role struct {
	ID        uuid.UUID `gorm:"type:uuid;primaryKey;default:uuid_generate_v4()"`
	Name      string    `gorm:"type:varchar(30);not null"`
	CreatedAt time.Time `gorm:"autoCreateTime"`
	UpdatedAt time.Time `gorm:"autoUpdateTime"`
}